requirements.txt
aicrowd-cli==0.1.15
certifi==2022.12.7
charset-normalizer==3.0.1
click==7.1.2
colorama==0.4.6
commonmark==0.9.1
gitdb==4.0.10
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
pip==21.1.2
psutil==5.9.4
Pygments==2.14.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-slugify==5.0.2
pyzmq==22.1.0
requests==2.28.2
requests-toolbelt==0.10.1
rich==10.16.2
scikit-learn==1.2.0
scipy==1.10.0
semver==2.13.0
smmap==5.0.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
torch==1.13.1
torch-geometric==2.2.0
torch-scatter==2.1.0
torch-sparse==0.6.16
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.14
